requirements.txt
pandas_redshift
pandas
python-decouple